foundation for Lobby folder structure refactoring
