--- src/features/GameList/TopNavbar/TopNavbar.css
@@ -1,3 +1,0 @@
-.topBar {
-
-}
