+++ src/features/Lobby/Lobby.css
@@ -13,32 +13,7 @@
     width: 70%;
     margin: 35px 0 35px 10px;
 }
-.gameList {
-    @media (--tablet) {
-        display: grid;
-        grid-template-columns: 50% 50%;
-        grid-template-rows: auto;
-    }
-}
 
-.gameListElement {
-    height: 60px;
-    border: 1px solid var(--game-list-elements-separator);
-    padding: 10px;
-    margin: 5px;
-    background-color: var(--game-list-element-bg);
-}
-
-.gameElementContainer {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    height: 100%;
-    border: 1px solid var(--game-list-element-border);
-    padding-left: 5px;
-    padding-right: 5px;
-    color: var(--game-list-element-font);
-}
 .secondColumn {
     margin-left: 10px;
     margin-right: 10px;
@@ -50,11 +25,6 @@
         width: 30%;
     }
 }
-.utilsRow {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-}
 
 .lobbyChat {
     position: relative;
@@ -64,3 +34,4 @@
     width: auto;
     flex-shrink: 0;
 }
+
