+++ src/features/Lobby/components/GameList/components/NoResults/NoResults.css
@@ -1,0 +1,9 @@
+.noResults {
+
+}
+
+.noResultsText  {
+
+
+}
+
