+++ src/features/Lobby/components/GameList/components/NoResults/NoResults.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+import styles from './NoResults.css'
+
+export const NoResults = () => (
+    <div className={styles.noResults}>
+        <span className={styles.noResultsText}>
+            srry no rooms
+        </span>
+    </div>
+)
