+++ src/features/Lobby/components/GameList/components/NoResults/index.js
@@ -1,0 +1,1 @@
+export { NoResults } from './NoResults'
