+++ src/features/Lobby/components/GameList/index.js
@@ -1,0 +1,1 @@
+export { default as GameList } from './GameList'
