+++ src/features/Lobby/components/PlayersList/PlayersListRow.js
@@ -1,10 +1,10 @@
 import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
 import classNames from 'classnames'
-import { getAvatar } from '../../../utils/avatarsHelper'
+import { getAvatar } from '../../../../utils/avatarsHelper'
 
 import styles from './PlayersList.css'
-import commonStyles from '../../Shared/CommonStyles/commonStyles.css'
+import commonStyles from '../../../Shared/CommonStyles/commonStyles.css'
 
 const PlayersListRow = ({ playerName, avatarNumber, currentRoom }) => {
     return (
