+++ src/features/Lobby/components/TopNavbar/TopNavbar.css
@@ -1,0 +1,5 @@
+.topBar {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+}
