+++ src/features/Lobby/components/TopNavbar/index.js
@@ -1,0 +1,1 @@
+export { default as TopNavbar } from './TopNavbar'
