+++ src/features/Lobby/index.js
@@ -1,0 +1,1 @@
+export { Lobby } from './Lobby'
